instance: Move instance.go under cmd package
All the capstan cmd is under the cmd package. Move it to the right
place.
